修改project_key_t
修改project_key_t的字段PJ_COST类型

+++ mysql_file/create.sql
@@ -203,7 +203,7 @@ CREATE TABLE if not exists project_key_t
 	PJ_XMMC varchar(80) comment '项目名称',
 	PJ_XMBH varchar(20) comment '项目编号',
 	PJ_KEY varchar(80) comment '俗称',
-	PJ_COST DECIMAL(5,4) comment '成本',
+	PJ_COST INTEGER comment '成本',
 	created_at datetime,
 	updated_at datetime
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
